+++ docs/HISTORY.rst
@@ -11,6 +11,10 @@ Changelog
   __ https://github.com/RedTurtle/redturtle.video/pull/14
 
 - Fix permission to add portlet. Now who can manage portlets can create it [cekk]
+- Added the uninstall profile (this close #12) [keul]
+
+  __ https://github.com/RedTurtle/redturtle.video/issues/12
+
 
 1.0.1 (2013-04-03)
 ------------------